refactoring: better separate components by folder
