+++ src/components/FavoritesView/FavoritesItem.svelte
@@ -1,8 +1,8 @@
 <script lang="ts">
   import type { FavoriteStore } from 'src/types/FavoritesStore';
-  import audioStreamGetter from '../lib/audioStreamGetter';
-  import { play, useSource, reset } from '../lib/AudioPlayer.svelte';
-  import IntersectionObserver from '../lib/IntersectionObserver.svelte';
+  import audioStreamGetter from '../../lib/audioStreamGetter';
+  import { play, useSource, reset } from '../../lib/AudioPlayer.svelte';
+  import IntersectionObserver from '../../lib/IntersectionObserver.svelte';
   import {
     musicTitle,
     poster,
@@ -11,11 +11,11 @@
     smallPoster,
     favorites,
     menuEntries,
-  } from '../lib/player';
+  } from '../../lib/player';
   import { fade } from 'svelte/transition';
 
   import truncate from 'just-truncate';
-  import binIcon from '../assets/bin.svg?raw';
+  import binIcon from '../../assets/bin.svg?raw';
 
   export let result: FavoriteStore;
   export let id: number;
