+++ src/components/FavoritesView/FavoritesList.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import audioStreamGetter from '../lib/audioStreamGetter';
-  import { play, useSource, reset, pause } from '../lib/AudioPlayer.svelte';
-  import ActionButton from '../lib/ActionButton.svelte';
+  import audioStreamGetter from '../../lib/audioStreamGetter';
+  import { play, useSource, reset, pause } from '../../lib/AudioPlayer.svelte';
+  import ActionButton from '../../lib/ActionButton.svelte';
 
   import {
     musicTitle,
@@ -12,15 +12,15 @@
     currentID,
     ended,
     smallPoster,
-  } from '../lib/player';
+  } from '../../lib/player';
   import { fade } from 'svelte/transition';
   import { flip } from 'svelte/animate';
 
-  import type { FavoriteStore } from '../types/FavoritesStore';
+  import type { FavoriteStore } from '../../types/FavoritesStore';
 
-  import Footer from './Footer.svelte';
+  import Footer from '../Footer.svelte';
   import FavoritesItem from './FavoritesItem.svelte';
-  import Modal from './Modal.svelte';
+  import Modal from '../../lib/Modal.svelte';
 
   let resultIndex = -1;
