+++ src/components/Toolbar/Autocomplete.svelte
@@ -3,7 +3,7 @@
 
   import { createEventDispatcher, onMount } from 'svelte';
   import truncate from 'just-truncate';
-  import { query } from '../lib/player';
+  import { query } from '../../lib/player';
 
   const dispatch = createEventDispatcher();
